some cleanup of code

+++ Assets/Scripts/Generic/WeightedObject.cs
@@ -6,7 +6,7 @@ public class WeightedObject
     public GameObject prefab;
     public float weight;
 
-    public WeightedObject(GameObject prefab, int weight)
+    public WeightedObject(GameObject prefab, float weight)
     {
         this.prefab = prefab;
         this.weight = weight;
